up to exercise 4.22

+++ part4/blog-list-app/controllers/blogs.js
@@ -1,29 +1,59 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1 })
+  
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
+
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+
+  console.log(`request token is ${request.token}`);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = new Blog ({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes ? body.likes : 0
+    likes: body.likes ? body.likes : 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+
+  await user.save()
+
   response.json(savedBlog.toJSON())
   
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  
+  const blog = await Blog.findById(request.params.id)
+  
+  const username = await User.findById(blog.user)
+
+
+  if (request.user === username.username) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(400).send({
+      error: 'unauthorized user'
+    })  }
 })
 
 //TODO: blogsRouter.delete is return 204 even when there wasn't any blog to delete
